requirements.txt

Flask==3.0.0
requests==2.31.0
PyYAML==6.0.1
pytest==8.2.2
pytest-asyncio==0.24.0
pytest-cov==4.1.0
coverage==7.3.2
ast-comments==1.1.2
